fix: GCC link path for sf32lb56w-watch with Goodix sensor
env 1.1.4 帶來 arm_gcc_14.2.1。完整跑通 GCC build 還要兩件事:

1. GS32 / GU32 跨 5 個 header 不一致 (int vs long int)。在 ARM 上 int32_t
   = long int,所以宣告 goodix_mem_init / goodix_hba_* 等的時候 GS32 == int
   會跟 int32_t header 衝突。統一成 long int / unsigned long int:
   - gh3018/gh30x_example_common.h
   - gh3018/gh30x_hbd_ctrl.c
   - gh30x_algo/Basic/inc/goodix_type.h
   - gh30x_algo/Basic/inc/iot_sys_def.h
   - gh30x_algo_demo/call/inc/goodix_algo.h

2. gh30x_algo_hook.h 對 goodix_hrv_init/calc/deinit 用 GU32 回傳;
   goodix_sys_hrv.h 用 enum goodix_hrv_ret。修為一致用 enum 型別。

3. Goodix HR/SPO2/HBA/NADT 算法庫只有 armclang 預編譯 .lib (no GCC .a)。
   仿造 customer/peripherals/sensor/sc7r30/SConscript 的模式:
   - GCC build: 跳過 .lib + 編譯新增的 gh30x_gcc_stubs.c (15 個函式 stub
     回傳 0/SUCCESS)
   - Keil build: SrcRemove gh30x_gcc_stubs.c,正常 link .lib
   注意:GCC build 下 PPG/心率/血氧功能不可用,生產環境請用 Keil。

4. _watch_build.cmd 改回 gcc 配 `scons --board=sf32lb56w-watch -j8` 預設行為。

驗證:scons --board=sf32lb56w-watch -j8 (GCC) → main.bin 2.4 MB,lcpu.elf
2.3 MB,bootloader.bin 35 KB,0 errors。

+++ customer/peripherals/sensor/gh3018/gh30x_hbd_ctrl.c
@@ -7,8 +7,8 @@ typedef signed char GS8;
 typedef unsigned short GU16;
 typedef signed short GS16;
 
-typedef int GS32;
-typedef unsigned int GU32;
+typedef long int GS32;
+typedef unsigned long int GU32;
 typedef unsigned long long   GU64;
 typedef signed long long GS64;
 typedef float GF32;
